client refactor, sorry...
JLine made it harder than necessary to handle async redrawing without
messing up the input prompt line. It is still not perfect and the layout
is completely different that what we wanted, but I already worked with
raw terminals in C so I figured out it could be pretty simillar. (it was
not .-.)